move SingleShowing.tsx into SingleShowing folder
