+++ src/AppRouter.tsx
@@ -1,7 +1,7 @@
 import { createBrowserRouter, RouterProvider } from "react-router-dom";
 import Showings from "./Showings/Showings";
 import Login from "./Login";
-import SingleShowing from "./SingleShowing";
+import SingleShowing from "./SingleShowing/SingleShowing";
 
 const router = createBrowserRouter([
     {
